Moves helper to demo app, not the addon itself. Updates relevant test.
